# docs/openapi/andy-policies-v1.yaml
openapi: 3.0.1
info:
  title: Andy Policies API
  description: 'Governance policy catalog — structured, versioned policy documents with lifecycle and audit trail, consumed by Conductor for story admission, verification, and compliance reporting (content only; enforcement lives in consumers).'
  contact:
    name: Rivoli AI
    url: https://github.com/rivoli-ai/andy-policies
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: v1
paths:
...
  /api/scopes:
    get:
      tags:
        - Scopes
      summary: "List scope nodes. Optional `?type=` filter narrows to a\r\nsingle ScopeType (Org/Tenant/Team/Repo/Template/Run); without\r\nthe filter the entire catalogue is returned ordered by\r\n(Depth, Ref)."
      operationId: Scopes_List
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ScopeType'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScopeNodeDto'
    post:
      tags:
        - Scopes
      summary: "Create a new scope node. Andy.Policies.Application.Dtos.CreateScopeNodeRequest.ParentId\r\nis null for a root Andy.Policies.Domain.Enums.ScopeType.Org node and\r\nrequired otherwise. The service enforces the canonical\r\nOrg→Tenant→Team→Repo→Template→Run ladder; mismatched type\r\nreturns 400."
      operationId: Scopes_Create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScopeNodeRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateScopeNodeRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateScopeNodeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeNodeDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  '/api/scopes/{id}':
    get:
      tags:
        - Scopes
      summary: Get a single scope node by id. Returns 404 if not found.
      operationId: Scopes_Get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeNodeDto'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Scopes
      summary: "Delete a leaf scope node. Refuses with 409 if the node still\r\nhas children; consumers must walk the subtree and delete\r\nleaves first."
      operationId: Scopes_Delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteScopeNodeRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/DeleteScopeNodeRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DeleteScopeNodeRequest'
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/scopes/{id}/effective-policies':
    get:
      tags:
        - Scopes
      summary: "Resolve the effective policy set for a scope node using the\r\nstricter-tightens-only fold (P4.3). Returns 404 if the node\r\ndoesn't exist."
      operationId: Scopes_Effective
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: bundleId
          in: query
          description: Bundle id to pin the read against. Required when andy.policies.bundleVersionPinning is true (the manifest default); absence yields a 400 ProblemDetails with type "https://andy.local/problems/bundle-pin-required".
          schema:
            type: string
            format: uuid
        - name: principalSubjectId
          in: query
          schema:
            type: string
        - name: cohort
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectivePolicySetDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          x-andy-pinning-gate-type: https://andy.local/problems/bundle-pin-required
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    AppliedOverrideDto:
      type: object
      properties:
        overrideId:
          type: string
          format: uuid
        effect:
          $ref: '#/components/schemas/OverrideEffect'
        scopeKind:
          $ref: '#/components/schemas/OverrideScopeKind'
        scopeRef:
          type: string
          nullable: true
        originalPolicyVersionId:
          type: string
          format: uuid
        replacementPolicyVersionId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
...
    BindStrength:
      enum:
        - Mandatory
        - Recommended
      type: string
...
    CreateScopeNodeRequest:
      type: object
      properties:
        parentId:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: '#/components/schemas/ScopeType'
        ref:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        rationale:
          type: string
          nullable: true
      additionalProperties: false
      description: "Request payload for `IScopeService.CreateAsync` (P4.2, story\r\nrivoli-ai/andy-policies#29). Andy.Policies.Application.Dtos.CreateScopeNodeRequest.ParentId is null for a\r\nroot Andy.Policies.Domain.Enums.ScopeType.Org node; non-null for any other type.\r\nThe service enforces the canonical Org→Tenant→Team→Repo→Template→Run\r\nladder."
...
    DeleteScopeNodeRequest:
      type: object
      properties:
        rationale:
          type: string
          nullable: true
      additionalProperties: false
      description: Audit rationale for deleting a leaf scope node.
    EffectivePolicyDto:
      type: object
      properties:
        policyId:
          type: string
          format: uuid
        policyVersionId:
          type: string
          format: uuid
        policyKey:
          type: string
          nullable: true
        version:
          type: integer
          format: int32
        bindStrength:
          $ref: '#/components/schemas/BindStrength'
        sourceBindingId:
          type: string
          format: uuid
        sourceScopeNodeId:
          type: string
          format: uuid
          nullable: true
        sourceScopeType:
          $ref: '#/components/schemas/ScopeType'
        sourceDepth:
          type: integer
          format: int32
      additionalProperties: false
      description: "One entry in an effective-policy set (P4.3, story\r\nrivoli-ai/andy-policies#30). The result of stricter-tightens-only\r\nresolution: each Andy.Policies.Application.Dtos.EffectivePolicyDto.PolicyId appears at most once,\r\ncarrying the strictest Andy.Policies.Application.Dtos.EffectivePolicyDto.BindStrength seen anywhere in\r\nthe scope chain plus a pointer back to the binding (and the scope\r\nnode that bound it) that won the resolution."
    EffectivePolicySetDto:
      type: object
      properties:
        scopeNodeId:
          type: string
          format: uuid
          nullable: true
        policies:
          type: array
          items:
            $ref: '#/components/schemas/EffectivePolicyDto'
          nullable: true
        appliedOverrides:
          type: array
          items:
            $ref: '#/components/schemas/AppliedOverrideDto'
          description: "Deterministic explanation of every override that changed\r\n            the baseline set, including Exempt entries whose policy is absent\r\n            from Andy.Policies.Application.Dtos.EffectivePolicySetDto.Policies."
          nullable: true
      additionalProperties: false
      description: "Envelope for `IBindingResolutionService` results (P4.3, story\r\nrivoli-ai/andy-policies#30). Andy.Policies.Application.Dtos.EffectivePolicySetDto.ScopeNodeId is null when\r\nthe resolver was called with a target that does not map to a known\r\n`ScopeNode` — the service then degrades to P3 exact-match\r\nsemantics rather than returning 404."
...
    OverrideEffect:
      enum:
        - Exempt
        - Replace
      type: string
    OverrideScopeKind:
      enum:
        - Principal
        - Cohort
      type: string
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    ScopeNodeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: '#/components/schemas/ScopeType'
        ref:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        materializedPath:
          type: string
          nullable: true
        depth:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      description: "Wire-format projection of a `ScopeNode` (P4.2, story\r\nrivoli-ai/andy-policies#29). Surface controllers (REST P4.5, MCP /\r\ngRPC / CLI in P4.6) emit this shape directly so wire behaviour stays\r\nuniform across surfaces."
...
    ScopeType:
      enum:
        - Org
        - Tenant
        - Team
        - Repo
        - Template
        - Run
      type: string
...
  securitySchemes:
    Bearer:
      type: http
      description: JWT Authorization header using the Bearer scheme.
      scheme: bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]
...